projects/titanic/requirements.txt
kaggle==1.6.8
python-dotenv==1.0.1
pre-commit==3.7.0
ydata-profiling==4.7.0
mlflow==2.11.3